Add category field for sarif report in daily scan
Update daily-scan.yaml to add the category field for the sarif report. We need this because we need a way to make sure we can differentiate the trivy filesystem and image scans separate from dependabot. It also should let me associate the results with a specific image/release with Vanta.